requirements.txt
flask-sqlalchemy
flask-babel==4.0.0
python-dotenv
psycopg2-binary
pytest
pytest-cov
codecov


